drivers/net/tun.c: added support for TAP multiframe mode
It's now possible to write to and read from a tap device more than
one frame with a single system call. The readv and writev system calls
have been modified to behave in this way when the TAP multiframe mode
is enabled. You can enable/disable the new modality issuing an ioctl

	ioctl(tapfd, TUNSETMULTIFRAME, value);

on the tap file descriptor, where value is 1 (enable) or 0 (disable).