tun.c: multiframe mode fix
When you issue a multiframe writev(), and there are multiple destinations
among the frames provided to the system call, you want to notify each
destination (tell them that the burst is done) at least once.
This patch get this requirement met, and tries to send only one notify
per destination with a svery straightforward algorithm.