e1000: CSB exit protocol fixed
It's important to execute the CSB exit protocol (writing 0xFFFFFFFF
to CSBBAH) at the very beginning of e1000_close(), so that the
synchronization with the TX BH happens before the TX/RX is
disabled in the hardware (TCTL/RCTL). Otherwise we would ask
for TX processing with TX disabled.

For these reasonse the CSB exit protocol has been moved at the beginning
of e1000_down() (which is called at the beginning of e1000_close()), just
before TX/RX are disabled in the hardware.