e1000: Complete paravirtualization
The CSB variables guest_need_*xkick are used instead of the IE register,
which now is not normally accessed. In this way we avoid two register access
per interrupt, and therefore increase the RX/TX throughput.

Because of the NAPI constraints, we cannot receive an interrupt from a device
while a napi context is scheduled/running (e.g. between napi_schedule() and
napi_complete()). We use the two CSB variables
as they were a single variable, because they must be 0 in the critical
section and we want them to be 1 outside the critical section.