rtl8139cp: paravirtual guest_need_txkick used
In paravirtual mode, guest_need_txkick is always zero when there
are free TX descriptors (e.g. when the kernel is allowed to send
frames through the rtl8139cp interface). The used TX descriptor are
cleaned at the beginning of the start_xmit() routine, so that TX
interrupts are not necessary. When we run out of TX descriptors
(e.g. when we tell the kernel to stop sending frames through
the interface), however, we set guest_need_txkick because we we won't
have any chance to clean new used TX descriptors in the start_xmit()
routine, and therefore we need TX interrupts. Once we make some new
free TX descriptors we can clear guest_need_txkick again.