8139cp.c: guest_request_txkick
The flag guest_need_txkick is not used anymore (is always zero).
Instead the field guest_request_txkick is used to request an
interrupt when the driver runs out of free TX descriptor.
In this way the interrupt can be delayed until the hypervisor
has processed many descriptor (and not just one).