e1000: IP checksumming in paravirtual mode
When in paravirtual mode and a TSO packet comes from the kernel,
the driver calls e1000_tso(), which used to zero the IP checksum field,
since it expects the emulator to compute the checksum. In turn, the
emulator expects the IP checksum to be zero.
However, this is silly since we are using the virtio-net header and so
we don't segment the TSO packet for real, and the guest kernel hsa already
computed the IP checksum on the TSO packet. We can just leave the
checksum as is (correct), and don't set E1000_TXD_POPTS_IXSM bit in the
associated TX descriptor, the emulator won't then avoid computing the
checksum.