e1000: tcp partial checksum computation
In paravirtual mode, with TSO packets, the driver used to compute
a partial checksum in e1000_tso(), using the csum_tcpudp_magic()
function. Once again (see the last commit), this is not necessary since
we are using virtio-net header, and so we can hand off an unchecksummed
packet (that won't be segmented) to the emulator.
This saves CPU cycles, incrementing the TCP throughput.