e1000: move back tcpudp_magic out of if() { }
This is necessary for now, so that TSO works (very slowly, but
works) with QEMU backends that do not support the virtio-net
header (we have decided to require the driver to ALWAYS use
the header in paravirtual mode).
In the future also the tcpudp_magic() operation will be done
by the QEMU frontend, so that we can move again this inside
the if construct.